ecommerce/order/serializers.py
from rest_framework import serializers
from rest_framework import status
from .models import *


class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = '__all__'
